# Speed up get_scalar_term for large vector constraints