Getting the set and function repeatedly for each index of the same constraint is slower than getting it once for each constraint.

On the benchmark of https://github.com/jump-dev/MathOptInterface.jl/issues/1796, in terms of allocations:

|                | theta1    | thetaG11   |
|----------------|-----------|------------|
| v0.5.2 | 14.693 MB | 793.687 GB |
| https://github.com/jump-dev/Dualization.jl/pull/135  | 0.395 MB | 0.039 GB   |
| https://github.com/jump-dev/Dualization.jl/pull/136  | 0.349 MB | 0.029 GB   |
| https://github.com/jump-dev/Dualization.jl/pull/137  | 0.349 MB | 0.029 GB   |

and in terms of time:

|                | theta1    | thetaG11   |
|----------------|-----------|------------|
| v0.5.2 | 28.5 ms | 232 s |
| https://github.com/jump-dev/Dualization.jl/pull/135 | 0.586 ms | 0.108 s   |
| https://github.com/jump-dev/Dualization.jl/pull/136 | 0.397 ms | 0.068 s |
| https://github.com/jump-dev/Dualization.jl/pull/137 | 0.246 ms | 0.029 s |